Add git-rebase-chain for rebasing multiple branches at once.
git-rebase-chain takes a destination branch and several source branches,
and will attempt to rebase the first source branch onto the destination,
then the next onto the new location of the previous branch, and so on.
It's useful when you have a number of branches on top of each other and
want to transfer the entire list into a new location.