[git-rebase-chain] Support --skip, --continue, and --abort.
When rebasing, the state for the rebase is now stored in a
.git/rebase-chain-state file. We can use this to resume an in-progress
rebase when the user needs to resolve conflicts.

These work just like git rebase's options, but will continue on with the
chain (in the case of skip/continue) once things are resolved.

This is still experimental and may need further tweaking.