[git-rebase-chain] Add short-hand for HEAD and ranges of branches.
You can now specify '.' when using the current branch as the
destination.

Instead of writing out every branch to rebase, you can now specify a
range using '..'. For example, 'branch1..branch10' will look up all
branches that are within that range (in the tree) and provide them, as
if they were typed individually.