[git-rebase-chain] Fix parent branch calculation in some cases.
Parent branches weren't properly used in all cases. If you were rebasing
a clean, linear history, and wanted to take part of it off the middle
and place it elsewhere, specifying the parent wouldn't end up working.
This was due to a bad calculation where the parent branch was being
overridden. We now sanity-check this better.