[git-j] Store aliases in .gitconfig files.
Aliases are now stored in the git config file instead of a custom
j-config file. This will allow us to support global aliases.

Since Git config keys are limited to alphanumeric characters and '-',
and must start with a letter, we can't use the alias as the key.
Instead, we now store keys in the following form:

    [git-j]
	    alias = <name>=<value>
	    alias = <name>=<value>

This is a bit annoying when hand-editing, but allows us to remain
flexible with alias names.